1.4.6 修正了预览
