+++ app/src/main/res/layout/activity_detail_player.xml
@@ -29,7 +29,7 @@
 
     </android.support.v4.widget.NestedScrollView>
 
-    <com.shuyu.gsyvideoplayer.video.StandardGSYVideoPlayer
+    <com.shuyu.gsyvideoplayer.video.CustomGSYVideoPlayer
         android:id="@+id/detail_player"
         android:layout_width="match_parent"
         android:layout_height="@dimen/post_media_height" />
