+++ gsyVideoPlayer/src/main/res/values/dimens.xml
@@ -4,4 +4,6 @@
     <dimen name="video_progress_dialog_margin_top">80dp</dimen>
     <dimen name="video_volume_dialog_margin_left">24dp</dimen>
 
+
+    <dimen name="seek_bar_image">150dp</dimen>
 </resources>
